+++ src/config.ts
@@ -2,8 +2,7 @@ export default {
   cssId: 'ja-contextmenu-css',
   panelClassName: 'ja-panel',
   wrapperClassName: 'ja-contextmenu',
-  defaultMenuWidth: 200, // edge: 250
-  // childMenuWidth: 150, // 自动继承父菜单宽度
+  defaultMenuWidth: 200,
   menuItemHeight: 24,
   menuItemDivideLineMargin: 5, // type == 'divide'
   defaultZIndex: 5000,