+++ src/index.ts
@@ -1,4 +1,4 @@
-// import 'core-js';
 export { default } from './ContextMenu';
 export { default as Panel } from './Panel';
 export { default as h } from './utils/h';
+// import './style.less';